fixed madgwick issues needs a little tuning still. added a socketio to redis adapter so the helper apps can be run on remote clients

+++ Ricardo_OS/Ricardo_OS/lib/librnp/src/rnp_networkmanager.cpp
@@ -121,9 +121,14 @@ void RnpNetworkManager::sendByRoute(const Route& route, RnpPacket& packet)
 
 void RnpNetworkManager::setAddress(uint8_t address) 
 {
-    routingtable.deleteRoute(_currentAddress);
+
+    auto currentRoute = routingtable.getRoute(_currentAddress);
+    if (currentRoute && (currentRoute.value().iface == static_cast<uint8_t>(DEFAULT_INTERFACES::LOOPBACK))){
+        routingtable.deleteRoute(_currentAddress); // ensure we dont delete a new route if this is called after a new routing table is assigned
+    }
     _currentAddress = address;
-    routingtable.setRoute(_currentAddress,Route{0,1,{}});
+    generateDefaultRoutes();
+   
 };
 
 void RnpNetworkManager::setNodeType(NODETYPE nodeType)
